refactor : 그룹 도메인 및 서비스 테스트 코드 분리, 오탈자 수정
